feat: add commands support and solve table creation error
Changed how the program is ran. It now supports a command interface to execute operations.
Only table and user creation are implemented yet.

Segregated the db object initialization and the server logic.

Added a catch block to the db.none promise which creates the table, and now the program
returns an error saying that the table was not created because it already exists.