fix: Address issues with code readibility
Changed the commandList enum to class. The class has a constructor with two arguments: the command input string and the instructions on how to use it. The class has all the commands as static instances of itself, and an override of the toString() method, to print the help menu entry.

The dispatchCommand function has now a return type of Promise<any> and the switche's default has been changed to also return a promise, as the program was throwing an exception when trying to execute the .then on a wrong command input.

Import statement for the server module functions on the dispatcher was also changed for readibility.

The data-processor module was changed to increase readibility.